requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
Django==5.1.6
django-phonenumber-field==8.0.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
phonenumbers==8.13.54
propcache==0.2.1
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyngrok==7.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
yarl==1.18.3
